backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==2.4.0
djangorestframework-simplejwt==4.8.0
asgiref==3.3.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
reportlab==3.6.3
sqlparse==0.3.1
python-dotenv==0.19.2
Pillow==9.0.1
djoser==2.1.0
webcolors==1.12
drf-base64==2.0
fpdf==1.7.2
isort==5.10.1
django-colorfield==0.8.0
sentry-sdk==1.25.0